Rename file to camel case
